py-be/pyproject.toml
[tool.poetry]
name = "py-be"
version = "0.1.0"
description = ""
authors = ["pulkit-khullar <pulkitkhullar15@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
